+++ src/solver.py
@@ -1,0 +1,5 @@
+from domain import *
+from map import Map
+
+def solve(map: Map) -> int:
+    while any(any (map.stores));
